requirements.txt
openai
tiktoken
pytest
nest_asyncio
requests
pydantic
python-dotenv
langchain
pytest-asyncio